chore(TS): setup redux for typescript
